requirements.txt
asgiref
Django
python-decouple
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
PyJWT
pytz
sqlparse
tzdata
